Replace 'exec -a' usage.
The -a switch on exec in bash sets the command that would show up in the
process table.  This string is used to exactly match the running sleep
process so it can be killed if necessary.  POSIX exec doesn't support
-a, so instead just use exec without the -a switch.  For the pkill
usage, match the basic string of 'sleep SECONDS' but also check the
parent process ID to make sure it matches our running script process.